+++ src/signals/signals/index.ts
@@ -1,5 +1,6 @@
-export * from "./State";
 export * from "./Computed";
 export * from "./Effect";
+export * from "./LocalSignal";
 export * from "./LocalState";
 export * from "./Signal";
+export * from "./State";